Endpoints return value

+++ src/test/java/com/wine/to/up/winelab/parser/service/dto/WineTest.java
@@ -61,14 +61,14 @@ class WineTest {
             ParserApi.Wine apiWine = wine.toParserWine();
 
             Assertions.assertEquals("Вино Saga Domaine Barons de Rothschild Bordeaux красное сухое 0,75 л", apiWine.getName());             //test the fields are being mapped correctly
-            Assertions.assertEquals((float) 1243.0, apiWine.getOldPrice());
+            Assertions.assertEquals(1243.0f, apiWine.getOldPrice());
             Assertions.assertEquals("https://www.winelab.ru/product/1014769", apiWine.getLink());
-            Assertions.assertEquals((float) 599.0, apiWine.getNewPrice());
+            Assertions.assertEquals(599.0f, apiWine.getNewPrice());
             Assertions.assertEquals("https://www.winelab.ru/medias/1014769.png-300Wx300H?context=bWFzdGVyfGltYWdlc3wzMzQwMXxpbWFnZS9wbmd8aW1hZ2VzL2g4NC9oZWMvODgzMjY0NjkzODY1NC5wbmd8OTk3MDg5NjdlMTk4NzlhNWM2MWQ0YzBiZGNhZmFmNGM3ZDViYmU1NWJmMzgyNDUwNWY0ZmRiYjczODdmOTJhOA", apiWine.getImage());
             Assertions.assertEquals("Domaine Barons de Rothschild", apiWine.getManufacturer());
             Assertions.assertEquals("SAGA", apiWine.getBrand());
             Assertions.assertEquals("Франция", apiWine.getCountry());
-            Assertions.assertEquals((float) 0.75, apiWine.getCapacity());
+            Assertions.assertEquals(0.75f, apiWine.getCapacity());
             Assertions.assertEquals(ParserApi.Wine.Color.RED, apiWine.getColor());
             Assertions.assertEquals(ParserApi.Wine.Sugar.DRY, apiWine.getSugar());
             Assertions.assertEquals("Каберне Совиньон", apiWine.getGrapeSort(0));
